+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## [1.4.0]
 
-* Removed `int WebSocketTimeout` and `int restTimeout` parameters and replaced them with a `Duration timeout` parameter.
+* Removed `int WebSocketTimeout` and `int restTimeout` parameters and replaced them with an `int timeout` parameter.
 * Added possibility to set a `receiveWindow` parameter that prevents from replay attacks
 * Minor doc changes